# 校园智能助手API-OpenAPI.yaml
openapi: 3.0.3
info:
  title: 校园智能助手 API
  description: |
    校园智能助手"小椰"的 RESTful API 接口文档。
    
    ## 功能说明
    - 简单对话接口（v1）：无状态，适合快速测试
    - 高级对话接口（v2）：支持多轮对话，具有上下文记忆功能
    
    ## Agent 能力
    - 查询天气信息
    - 查询课程表
    - 查询图书馆座位
    - 回答校园生活问题
    
  version: 1.0.0
  contact:
    name: 开发团队

servers:
  - url: http://localhost:8082
    description: 开发环境

tags:
  - name: 简单对话
    description: 无状态的简单对话接口
  - name: 高级对话
    description: 支持多轮对话的高级接口

paths:
  /api/v1/chat:
    get:
      tags:
        - 简单对话
      summary: 简单对话接口
      description: |
        简单的单次对话接口，不支持上下文记忆。
        每次请求都是独立的，适合快速测试和简单查询。
      operationId: simpleChat
      parameters:
        - name: question
          in: query
          description: 用户的问题
          required: false
          schema:
            type: string
            default: "图书馆在哪里"
            example: "今天天气怎么样"
      responses:
        '200':
          description: 成功返回 AI 回答
          content:
            text/plain:
              schema:
                type: string
                example: "今天南京的天气是晴天，温度22°C，适合出行，建议穿长袖。"
        '500':
          description: 服务器内部错误

  /api/v2/chat:
    post:
      tags:
        - 高级对话
      summary: 高级对话接口
      description: |
        支持多轮对话的智能对话接口，具有上下文记忆功能。
        后端会为每个 userId 维护一个 threadId，实现对话连续性。
        
        ## 多轮对话示例
        1. 第一轮：用户说"我是学号20220001的学生"
        2. 第二轮：用户说"帮我查一下我的课程表"
        3. AI 会记住之前提到的学号，自动关联上下文
      operationId: advancedChat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestDTO'
            examples:
              basic:
                summary: 基本示例
                value:
                  userId: "user123"
                  message: "帮我查一下今天的天气"
              multiTurn:
                summary: 多轮对话示例
                value:
                  userId: "user123"
                  message: "我是学号20220001的学生"
      responses:
        '200':
          description: 成功返回 AI 回答
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssistantResponse'
              examples:
                weather:
                  summary: 天气查询响应
                  value:
                    userId: "user123"
                    threadId: "550e8400-e29b-41d4-a716-446655440000"
                    answer: "今天南京的天气是晴天，温度22°C，适合出行，建议穿长袖。"
                    type: "weather"
                    suggestion: "建议根据天气情况选择合适的衣物"
                    needsFurtherHelp: false
                course:
                  summary: 课程查询响应
                  value:
                    userId: "user123"
                    threadId: "550e8400-e29b-41d4-a716-446655440000"
                    answer: "根据你的学号20220001，你的课程安排如下：\n1. Spring Boot - 08:00-09:40 - 教四319\n2. Spring AI - 13:30-15:10 - 教四419"
                    type: "course"
                    suggestion: "记得提前到教室哦"
                    needsFurtherHelp: false
        '500':
          description: 服务器内部错误

  /api/v2/history/{userId}:
    get:
      tags:
        - 高级对话
      summary: 查询对话历史
      description: |
        根据 userId 查询该用户对应的对话 threadId 和简化的历史信息。
        目前历史内容为占位实现，实际历史记录功能需要进一步开发。
      operationId: getHistory
      parameters:
        - name: userId
          in: path
          required: true
          description: 用户ID
          schema:
            type: string
            example: "user123"
      responses:
        '200':
          description: 成功返回历史信息
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/HistoryResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
              examples:
                success:
                  summary: 成功响应
                  value:
                    userId: "user123"
                    threadId: "550e8400-e29b-41d4-a716-446655440000"
                    history: "历史记录功能需要进一步实现"
                error:
                  summary: 错误响应
                  value:
                    error: "未找到该用户的历史记录"
        '500':
          description: 服务器内部错误

components:
  schemas:
    RequestDTO:
      type: object
      required:
        - userId
        - message
      properties:
        userId:
          type: string
          description: 用户唯一标识，用于生成和维护 threadId。相同 userId 的请求会共享同一个对话上下文
          example: "user123"
        message:
          type: string
          description: 用户发送的消息内容
          example: "帮我查一下今天的天气"
      example:
        userId: "user123"
        message: "帮我查一下今天的天气"

    AssistantResponse:
      type: object
      required:
        - answer
        - type
        - suggestion
        - needsFurtherHelp
      properties:
        userId:
          type: string
          description: 用户ID
          example: "user123"
        threadId:
          type: string
          format: uuid
          description: 线程ID，用于维护对话上下文
          example: "550e8400-e29b-41d4-a716-446655440000"
        answer:
          type: string
          description: AI 返回的主要回答内容（自然语言）
          example: "今天南京的天气是晴天，温度22°C，适合出行，建议穿长袖。"
        type:
          type: string
          enum:
            - weather
            - course
            - library
            - general
          description: 回答类型
          default: general
          example: "weather"
        suggestion:
          type: string
          description: 给用户的建议文案
          default: "请根据你的需求，给出一个建议"
          example: "建议根据天气情况选择合适的衣物"
        needsFurtherHelp:
          type: boolean
          description: 是否需要进一步帮助
          default: false
          example: false
      example:
        userId: "user123"
        threadId: "550e8400-e29b-41d4-a716-446655440000"
        answer: "今天南京的天气是晴天，温度22°C，适合出行，建议穿长袖。"
        type: "weather"
        suggestion: "建议根据天气情况选择合适的衣物"
        needsFurtherHelp: false

    HistoryResponse:
      type: object
      required:
        - userId
        - threadId
        - history
      properties:
        userId:
          type: string
          description: 用户ID
          example: "user123"
        threadId:
          type: string
          format: uuid
          description: 线程ID
          example: "550e8400-e29b-41d4-a716-446655440000"
        history:
          type: string
          description: 历史记录内容（当前为占位文本）
          example: "历史记录功能需要进一步实现"

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: 错误信息
          example: "未找到该用户的历史记录"

  examples:
    MultiTurnConversation:
      description: 多轮对话示例
      value:
        turn1:
          request:
            userId: "user123"
            message: "我是学号20220001的学生"
          response:
            answer: "你好！我是小椰，很高兴为你服务。"
            threadId: "xxx-xxx-xxx"
        turn2:
          request:
            userId: "user123"
            message: "帮我查一下我的课程表"
          response:
            answer: "根据你的学号20220001，你的课程安排如下：..."
            threadId: "xxx-xxx-xxx"
          note: "AI 会记住之前提到的学号，因为使用了相同的 userId 和 threadId"